add udacity media query course link
Hi again,

This is an extremely high quality (and free) course provided by Udacity and Google. This course has 74 pages which are either videos or quizzes. It teaches creating custom media queries and what responsive design means at a high level, the importance of making content work across any device so as not to lose users.